requirements.txt
Flask==3.0.0
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
requests==2.31.0
apscheduler==3.10.4
